First commit 1/2
Sending working files using 1.8.1 pokeemerald expansion with some added maps and new/edited scripts

+++ src/debug.c
@@ -3404,7 +3404,7 @@ static void DebugAction_Give_Pokemon_SelectNature(u8 taskId)
 
     if (JOY_NEW(A_BUTTON))
     {
-        u16 abilityId;
+        u8 abilityId;
         sDebugMonData->nature = gTasks[taskId].tInput;
         gTasks[taskId].tInput = 0;
         gTasks[taskId].tDigit = 0;
